requirements.txt
pandas==2.3.3
numpy==2.3.4
torch==2.9.0
matplotlib==3.10.7
pytest==8.4.2
hdbscan==0.8.40
PyYAML==6.0.3
scikit-learn==1.7.2
seaborn==0.13.2
networkx==3.5